net/rpmsg: fix possible kernel crash in callback for deleted sockets
The rpmsg proto driver creates a rpmsg channel device per remote
processor for each MessageQ object through the socket's bind()
call (a Rx socket) for receiving messages. During error recovery,
the virtio core removes the devices created by the Rx sockets,
while the open sockets themselves can be freed up anytime either
by the userspace closing them or through an automatic close when
the process is terminated/closed. Commit e5a9a772354e ("net/rpmsg:
set error status on open sockets after recovery") added the support
for error recovery and handles the race conditions between both the
rpmsg_sock_release() and rpmsg_proto_remove() paths.

The rpmsg_sock_release() clears the end-point's private field for Rx
sockets before destroying the endpoint. This presents a small window
(until before the __rpmsg_destroy_ept() is invoked in virtio rpmsg
core) during which a message can come in for a socket that has already
been deleted, and results in a NULL pointer dereference kernel panic
when the invalid (NULL) sk is accessed. This could happen, for example,
when an application has incorrectly terminated while still communicating
with the remote core. The remote core is unaware of the application's
termination and continues to send messages.

Fix this by adding a check in the rpmsg_proto callback for a NULL
(invalid) sk before accessing the socket, and throw a warning if a
message is received for any invalid socket. Note that this will not
notify the remote processor of the error and it may still be unaware
of the removal of the HOST rpmsg-proto socket unless it has some
other mechanism for checking that the socket is still alive. All
such messages are either dropped in the virtio rpmsg core or in the
rpmsg proto driver's callback.

Fixes: e5a9a772354e ("net/rpmsg: set error status on open sockets after recovery")
Signed-off-by: Angela Stegmaier <angelabaker@ti.com>
[s-anna@ti.com: add some more details to commit description]
Signed-off-by: Suman Anna <s-anna@ti.com>

Conflicts:
	net/rpmsg/rpmsg_proto.c

Change-Id: I76341ec2da8557db208746f2e1d798e241f890d7