arm: dts: dra7: Re-route PCF interrupts
For DRA72x EVM, the PCF irq gets fired as soon as the driver requests for irq handler.
This causes a NULL pointer crash and backtrace in printed.

Curerntly the IRQ is not used for anything.
Avoid this by removing the inteerupt related properties from PCF node.

Use a dummy-gpio line for tpd12s015 which provide hotplug support instead of
pcf_hdmi as the latter is disabled from linux.
Currently supplying pcf_lcd as the dummy-gpio line for tpd12s015.

Signed-off-by: Nikhil Devshatwar <nikhil.nd@ti.com>