net/rpmsg: skip announcement of rpmsg devices created in bind()
The rpmsg proto driver creates a rpmsg channel device per remote
processor (a Rx socket) for each MessageQ object through the
socket's bind() call. These rpmsg channel devices are normally
announced to the remote processor core, as these channels are
treated as rpmsg server channels. The behavior of announcing
the bound endpoints to the remote cores isn't necessary since
binding is for local endpoints. Furthermore, the TI MessageQ
stack doesn't require these devices to be announced as the
object will be looked up using a higher-level protocol logic.

Skip the announcing of these devices by overriding the rpmsg
device's announce field in probe, so that the subsequent
announce is skipped in the virtio rpmsg core. This approach
limits the customization to the rpmsg proto driver leaving
the virtio rpmsg core unchanged. It also has the added benefit
of removing the extra delay of sending the notification to
the remote processor.

Signed-off-by: Angela Stegmaier <angelabaker@ti.com>
[s-anna@ti.com: revise commit description]
Signed-off-by: Suman Anna <s-anna@ti.com>

Conflicts:
	net/rpmsg/rpmsg_proto.c

Change-Id: I44c6c969622ff7d00ef4f22e2131cdac08183154