media: ti-vpe: cal: Add per platform data support
commit a16e5cffeb5c ("media: ti-vpe: cal: Add per platform data support") branch ti-linux-4.9.y

First this patch adds a method to access the
CTRL_CORE_CAMERRX_CONTROL register to use the syscon
mechanism. For backward compatibility we also handle
using the existing camerrx_control "reg" entry if a syscon
node is not found.

In addition the register bit layout for the CTRL_CORE_CAMERRX_CONTROL
changes depending on the device. In order to support this we need
to use a register access scheme based on data configuration instead
of using static macro.

In this case we make use of the regmap facility and create
data set based on the various device and phy available.

Finally we update the devicetree binding.

Change-Id: I925386d1ff1417a006c94bc687c36031acf7630d
Signed-off-by: Benoit Parrot <bparrot@ti.com>
Signed-off-by: Jean-Jacques Hiblot <jjhiblot@ti.com>