clk: ti: convert to use proper register definition for all accesses
commit 6c0afb503937a12a8d20a805fcf263e31afa9871 upstream.

Currently, TI clock driver uses an encapsulated struct that is cast into
a void pointer to store all register addresses. This can be considered
as rather nasty hackery, and prevents from expanding the register
address field also. Instead, replace all the code to use proper struct
in place for this, which contains all the previously used data.

This patch is rather large as it is touching multiple files, but this
can't be split up as we need to avoid any boot breakage.

Acked-by: Tony Lindgren <tony@atomide.com>
[t-kristo@ti.com: fixed the dpll3xxx save/restore while backporting]
Signed-off-by: Tero Kristo <t-kristo@ti.com>

Signed-off-by: Jean-Jacques Hiblot <jjhiblot@ti.com>